moving to mysql

+++ reactive-data/src/main/java/com/fnbo/reactivedata/CommentRepository.java
@@ -2,7 +2,8 @@ package com.fnbo.reactivedata;
 
 import com.fnbo.reactivedata.model.Comment;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
-public interface CommentRepository extends CrudRepository<Comment, Integer> {
+public interface CommentRepository extends CrudRepository<Comment, Long> {
 
 }
